refactor: include code to fetch stats inside StatsMutation
This are the initial changes that will cause mutations to behave in a
transactionl-basis, where all reads are done inside the mutation
itself.
